the function leonie_failure_cleanup asks for a key to be pressed only if the pause mode is enabled
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@17269 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ leonie/source/leo_status.c
@@ -200,31 +200,36 @@ leo_help(void)
 }
 
 void
-leonie_failure_cleanup(void)
+leonie_failure_cleanup(p_leo_settings_t settings)
 {
   pid_t pid = 0;
 
   LOG_IN();
-  DISP("A failure has been detected, hit return to close the session");
-  getchar();
+  if (settings && settings->pause_mode) {
+    DISP("A failure has been detected, hit return to close the session");
+    getchar();
+  }
+  else {
+    DISP("A failure has been detected, killing process ...");
+  }
   pid = getpid();
   kill(-pid, SIGTERM);
   LOG_OUT();
 }
 
 void
-leo_terminate(const char *msg)
+leo_terminate(const char *msg, p_leo_settings_t settings)
 {
   DISP("error: %s", msg);
-  leonie_failure_cleanup();
+  leonie_failure_cleanup(settings);
   exit(EXIT_FAILURE);
 }
 
 void
-leo_error(const char *command)
+leo_error(const char *command, p_leo_settings_t settings)
 {
   perror(command);
-  leonie_failure_cleanup();
+  leonie_failure_cleanup(settings);
   exit(EXIT_FAILURE);
 }
